dachuang_project/dictionary_lookup.py
import requests

# 示例：使用一个在线API来查询词典
# 这里使用的是一个虚构的API，你需要替换成一个真实可用的API
API_URL = "https://api.dictionary.com/v3/references/collegiate/json/{word}?key=YOUR_API_KEY"

def lookup(word):
    """
    查询词典以获取词语的定义、拼音等信息。
    :param word: 要查询的词语
    :return: 词语的信息，或者在找不到时返回None
    """
    print(f"正在查询词语: {word}")
    # 在这里实现具体的查字典逻辑
    # 例如，可以调用一个在线词典API，或者查询本地词典数据库

    # --- 示例：本地模拟 ---
    mock_dictionary = {
        "苹果": "一种水果。",
        "香蕉": "一种黄色的水果。",
        "你好": "打招呼的常用语。"
    }
    return mock_dictionary.get(word, f"对不起，没有找到关于“{word}”的信息。")

if __name__ == '__main__':
    # 用于直接测试该模块
    test_word = "苹果"
    result = lookup(test_word)
    print(f"查询“{test_word}”的结果是: {result}")

    test_word = "电脑"
    result = lookup(test_word)
    print(f"查询“{test_word}”的结果是: {result}")
