dachuang_project/requirements.txt
pypinyin
pygame
webrtcvad-wheels
pyaudio
torch
openai-whisper
numpy
ultralytics
opencv-python
jieba
pydub